+++ Source/System/FileIO/3DS/OSG3DSLoader.cpp
@@ -748,7 +748,7 @@ void LMesh::CalcTextureSpace()
     // now for each vertex build a list of face that share this vertex
     std::vector< std::vector<int> > array;
     array.resize(m_vertices.size());
-    for (int i=0; i<m_triangles.size(); i++)
+    for (size_t i=0; i<m_triangles.size(); i++)
     {
         uint k = m_tris[i].a;
         array[k].push_back(i);
@@ -761,7 +761,7 @@ void LMesh::CalcTextureSpace()
     }
 
     // now average the tangents and compute the binormals as (tangent X normal)
-    for (int i=0; i<m_vertices.size(); i++)
+    for (size_t i=0; i<m_vertices.size(); i++)
     {
         v1 = zero3;
         v2 = zero3;
